feat: Add move number to critical moment challenge message
Updated the challenge prompt to include the full move notation with
move number and ellipsis when appropriate:

Before: "Nd4 played. Try a better move for white."
After:  "10.Nd4 played. Try a better move for white."

For Black moves: "12...Qc7 played. Try a better move for black."

This makes it clearer which move in the game the user is trying to
improve upon.